Factored out ChatNode class
